Reorganize files under packages
