+++ app/src/androidTest/java/com/example/cattlelog/UserFieldsDaoInstrumentedTest.kt
@@ -3,8 +3,8 @@ package com.example.cattlelog
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.example.cattlelog.dao.UserFieldsDao
-import com.example.cattlelog.database.CattlelogDatabase
+import com.example.cattlelog.model.dao.UserFieldsDao
+import com.example.cattlelog.model.database.CattlelogDatabase
 import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
