+++ app/src/main/java/com/example/cattlelog/adapter/CattleListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.cattlelog
+package com.example.cattlelog.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -6,7 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.cattlelog.entities.Cattle
+import com.example.cattlelog.R
+import com.example.cattlelog.model.entities.Cattle
 
 class CattleListAdapter internal constructor(
     context: Context
