+++ app/src/main/java/com/example/cattlelog/model/dao/TreatmentDao.kt
@@ -1,8 +1,8 @@
-package com.example.cattlelog.dao
+package com.example.cattlelog.model.dao
 
 import androidx.room.Dao
 import androidx.room.Query
-import com.example.cattlelog.entities.Treatment
+import com.example.cattlelog.model.entities.Treatment
 
 @Dao
 interface TreatmentDao {
