+++ app/src/main/java/com/example/cattlelog/model/repository/CattleRepository.kt
@@ -1,8 +1,8 @@
-package com.example.cattlelog
+package com.example.cattlelog.model.repository
 
 import androidx.lifecycle.LiveData
-import com.example.cattlelog.dao.CattleDao
-import com.example.cattlelog.entities.Cattle
+import com.example.cattlelog.model.dao.CattleDao
+import com.example.cattlelog.model.entities.Cattle
 
 // Declares the DAO as a private property in the constructor. Pass in the DAO
 // instead of the whole database, because you only need access to the DAO
@@ -11,5 +11,4 @@ class CattleRepository(private val cattleDao: CattleDao) {
     // Room executes all queries on a separate thread.
     // Observed LiveData will notify the observer when the data has changed.
     val allCattle: LiveData<List<Cattle>> = cattleDao.getAllCattle()
-
 }
